% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binaryNumber.R
\name{binaryNumber}
\alias{binaryNumber}
\title{Change Binary Number}
\usage{
binaryNumber(a)
}
\description{
This function that enters an natural number and changes it to binary number.
}
\examples{

binaryNumber(50)
}
